+++ RMSD/rmsd.cuh
@@ -4,6 +4,6 @@
 #include "point.h"
 
 
-double compute_rmsd_gpu(std::vector<point*> s1, std::vector<point*> s2);
+float compute_rmsd_gpu(std::vector<point*> s1, std::vector<point*> s2);
 
-__global__ void sumPointDistancesGPU(int* s1x, int* s1y, int* s1z, int* s2x, int* s2y, int* s2z, int* r);
+__global__ void sumPointDistancesGPU(float* s1x, float* s1y, float* s1z, float* s2x, float* s2y, float* s2z, float* r);
